# revert AESNI padding and handle the case in aes.c