Tested if this fix hurt performance by modifying benchmark.c and getting the average cycle/byte from 300 iterations on a tight loop of AES-GCM encrypt.

Without change
AES-GCM  50 megs took 0.053 seconds,  948.912 MB/s Cycles per byte =   2.60

With change
AES-GCM  50 megs took 0.052 seconds,  957.875 MB/s Cycles per byte =   2.59